Refactored exports to be named exports in named files
